requirements.txt
fastapi
uvicorn
python-dotenv
python-binance
